requirements.txt
# Core dependencies
pandas>=1.3.0
numpy>=1.20.0
flask>=2.0.0
werkzeug>=2.0.0

# File parsing
python-magic-bin>=0.4.14  # For Windows
# python-magic>=0.4.24    # For Linux/Mac
chardet>=4.0.0
pyyaml>=6.0
openpyxl>=3.0.9  # For Excel support
tabula-py>=2.3.0  # For PDF table extraction
pdfplumber>=0.7.0  # Alternative PDF parser
pytesseract>=0.3.8  # For OCR in PDFs

# Machine learning
torch>=1.10.0
transformers>=4.15.0

# Utilities
tqdm>=4.62.0
